helpit-analysis/requirements.txt
pymongo
requests
yampy3==1.0